LayoutTests/platform/mac-leopard/http/tests/inspector/console-websocket-error-expected.txt

CONSOLE MESSAGE: line 27: testDNSLookup: Test started.
CONSOLE MESSAGE: line 36: testDNSLookup: onclose is called.
CONSOLE MESSAGE: line 43: testSSLCertificate: Test started.
CONSOLE MESSAGE: line 55: testSSLCertificate: onclose is called.
WebSocket's network errors should be logged to console.

console-websocket-error.html:27testDNSLookup: Test started.
ws://nonexistent.domain.invalid:80/WebSocket network error: Operation could not be completed. (kCFErrorDomainCFNetwork error 2.)
console-websocket-error.html:36testDNSLookup: onclose is called.
console-websocket-error.html:43testSSLCertificate: Test started.
wss://127.0.0.1:8443/WebSocket network error: OSStatus Error -9812: valid cert chain, untrusted root 
console-websocket-error.html:55testSSLCertificate: onclose is called.
